Move logic into HWCleanup
