+++ include/circt/Dialect/SV/SVPasses.h
@@ -30,7 +30,6 @@ createHWMemSimImplPass(bool replSeqMem = false,
 std::unique_ptr<mlir::Pass> createSVExtractTestCodePass();
 std::unique_ptr<mlir::Pass>
 createHWExportModuleHierarchyPass(llvm::Optional<std::string> directory = {});
-std::unique_ptr<mlir::Pass> createHWGenerateCasePass();
 /// Generate the code for registering passes.
 #define GEN_PASS_REGISTRATION
 #include "circt/Dialect/SV/SVPasses.h.inc"
