+++ lib/Dialect/SV/Transforms/CMakeLists.txt
@@ -1,7 +1,6 @@
 add_circt_dialect_library(CIRCTSVTransforms
   GeneratorCallout.cpp
   HWCleanup.cpp
-  HWGenerateCase.cpp
   HWStubExternalModules.cpp
   HWLegalizeModules.cpp
   HWMemSimImpl.cpp
