+++ tools/firtool/firtool.cpp
@@ -569,7 +569,6 @@ processBuffer(MLIRContext &context, TimingScope &ts, llvm::SourceMgr &sourceMgr,
         modulePM.addPass(createCSEPass());
         modulePM.addPass(createSimpleCanonicalizerPass());
         modulePM.addPass(createCSEPass());
-        modulePM.addPass(sv::createHWGenerateCasePass());
         modulePM.addPass(sv::createHWCleanupPass());
       }
     }
